requirements.txt
Django==1.5.2
Markdown==2.3.1
bzr==2.5.1
dj-database-url==0.2.2
dj-static==0.0.5
django-extensions==1.1.1
django-toolbelt==0.0.1
djangorestframework==2.3.7
gunicorn==17.5
httplib2==0.8
psycopg2==2.5.1
python-dateutil==2.1
requests==1.2.3
six==1.3.0
static==0.4
wsgiref==0.1.2
